Use huggingface accelerate library in ddpm training and sampling

+++ utils/misc.py
@@ -1,34 +1,8 @@
 import os
 import sys
-import yaml
-import random
+import torch
 import shutil
 import datetime
-import numpy as np
-from typing import Dict
-from argparse import ArgumentParser
-
-import torch
-import torch.nn as nn
-from torch.backends import cudnn
-from torch.nn.parallel import DistributedDataParallel as DDP
-
-from utils.dist import main_process_only
-
-
-def init_seeds(seed: int = 0, cuda_deterministic: bool = False):
-    random.seed(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-    torch.cuda.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-    # Speed-reproducibility tradeoff https://pytorch.org/docs/stable/notes/randomness.html
-    if cuda_deterministic:  # slower, more reproducible
-        cudnn.deterministic = True
-        cudnn.benchmark = False
-    else:  # faster, less reproducible
-        cudnn.deterministic = False
-        cudnn.benchmark = True
 
 
 def check_freq(freq: int, step: int):
@@ -41,93 +15,64 @@ def get_time_str():
 
 
 def image_float_to_uint8(image: torch.Tensor):
+    """ [0, 1] -> [0, 255] """
     assert image.dtype == torch.float32
     assert (0 <= image).all() and (image <= 1).all()
     return (image * 255).to(dtype=torch.uint8)
 
 
 def image_norm_to_float(image: torch.Tensor):
+    """ [-1, 1] -> [0, 1] """
     assert image.dtype == torch.float32
     assert (-1 <= image).all() and (image <= 1).all()
     return (image + 1) / 2
 
 
 def image_norm_to_uint8(image: torch.Tensor):
+    """ [-1, 1] -> [0, 255] """
     assert image.dtype == torch.float32
     assert (-1 <= image).all() and (image <= 1).all()
     return ((image + 1) / 2 * 255).to(dtype=torch.uint8)
 
 
-def get_bare_model(model: nn.Module or DDP):
-    return model.module if isinstance(model, (nn.DataParallel, DDP)) else model
-
-
-def add_args(parser: ArgumentParser, cfg: Dict, prefix: str = ''):
-    """ Adapted from https://github.com/open-mmlab/mmcv/blob/master/mmcv/utils/config.py#L53 """
-    for k, v in cfg.items():
-        if isinstance(v, str):
-            parser.add_argument('--' + prefix + k, default=v)
-        elif isinstance(v, int):
-            parser.add_argument('--' + prefix + k, type=int, default=v)
-        elif isinstance(v, float):
-            parser.add_argument('--' + prefix + k, type=float, default=v)
-        elif isinstance(v, bool):
-            parser.add_argument('--' + prefix + k, type=bool, default=v)
-        elif isinstance(v, dict):
-            add_args(parser, v, prefix + k + '_')
-        elif isinstance(v, (list, tuple)):
-            parser.add_argument('--' + prefix + k, type=type(v[0]), nargs='+', default=v)
-        elif v is None:
-            parser.add_argument('--' + prefix + k, default=None)
-        else:
-            print(f'cannot parse key {prefix + k} of type {type(v)}')
-    return parser
-
-
 def find_resume_checkpoint(exp_dir: str, resume: str):
-    """ Checkpoints are named after 'stepxxxxxx.pt' """
-    if os.path.isfile(resume):
+    """ Checkpoints are named after 'stepxxxxxx/' """
+    if os.path.isdir(resume):
         ckpt_path = resume
     elif resume == 'best':
-        ckpt_path = os.path.join(exp_dir, 'ckpt', 'best.pt')
+        ckpt_path = os.path.join(exp_dir, 'ckpt', 'best')
     elif resume == 'latest':
         d = dict()
-        for filename in os.listdir(os.path.join(exp_dir, 'ckpt')):
-            name, ext = os.path.splitext(filename)
-            if ext == '.pt' and name[:4] == 'step':
-                d.update({int(name[4:]): filename})
+        for name in os.listdir(os.path.join(exp_dir, 'ckpt')):
+            if os.path.isdir(os.path.join(exp_dir, 'ckpt', name)) and name[:4] == 'step':
+                d.update({int(name[4:]): name})
         ckpt_path = os.path.join(exp_dir, 'ckpt', d[sorted(d)[-1]])
     else:
         raise ValueError(f'resume option {resume} is invalid')
-    assert os.path.isfile(ckpt_path), f'{ckpt_path} is not a .pt file'
+    assert os.path.isdir(ckpt_path), f'{ckpt_path} is not a directory'
     return ckpt_path
 
 
-@main_process_only
-def create_exp_dir(cfg_dict: Dict, resume: bool = False, time_str: str = None,
-                   name: str = None, no_interaction: bool = False):
+def create_exp_dir(
+        exp_dir: str,
+        cfg_dump: str,
+        exist_ok: bool = False,
+        time_str: str = None,
+        no_interaction: bool = False,
+):
     if time_str is None:
         time_str = get_time_str()
-    if name is None:
-        name = f'exp-{time_str}'
-    exp_dir = os.path.join('runs', name)
-    if os.path.exists(exp_dir) and not resume:
-        cover = True
-        if not no_interaction:
-            cover = query_yes_no(
-                question=f'{exp_dir} already exists! Cover it anyway?',
-                default='no',
-            )
-        if cover:
-            shutil.rmtree(exp_dir, ignore_errors=True)
-        else:
-            sys.exit(1)
+    if os.path.exists(exp_dir) and not exist_ok:
+        cover = no_interaction or query_yes_no(
+            question=f'{exp_dir} already exists! Cover it anyway?',
+            default='no',
+        )
+        shutil.rmtree(exp_dir, ignore_errors=True) if cover else sys.exit(1)
     os.makedirs(exp_dir, exist_ok=True)
     os.makedirs(os.path.join(exp_dir, 'ckpt'), exist_ok=True)
     os.makedirs(os.path.join(exp_dir, 'samples'), exist_ok=True)
     with open(os.path.join(exp_dir, f'config-{time_str}.yaml'), 'w') as f:
-        yaml.safe_dump(cfg_dict, f, sort_keys=False)
-    return exp_dir
+        f.write(cfg_dump)
 
 
 def query_yes_no(question: str, default: str = "yes"):
